Active data and code to generate it
Active data set is a JSON file which contains all of the origional data ('CharityCharacteristics.csv') for years 2011-12  meaning it contains 149 variables  (1 origional variable, ccnum is the key/index). 'Read_csv_to_JSON_tw.py' is the code used to generate this data. This data does not yet inclide up-to-date financials or Twitter informaiton